[core][cgraph] Split test_accelerated_dag (#52580)
